image upload, too many routes, picking right packages, configuration and more

+++ src/controllers/item.controller.ts
@@ -1,15 +1,156 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
-import { prisma } from '../utils';
-import { ERRORS, handleServerError } from '../helpers/errors.helper';
+import { prisma, utils } from '../utils';
+import { AppError, handleServerError } from '../helpers/errors.helper';
 import { STANDARD } from '../constants/request';
 import {
   CreateItemSchema,
+  ItemIDRequestParamSchema,
   ItemResponseSchema,
   UpdateItemSchema,
 } from '../schemas/Item';
 import { ISecureRequest } from '../types';
-import { User } from '@prisma/client';
-import { IRequestIdParamSchema } from '../schemas/Utils';
+import {
+  IRequestIdParamSchema,
+  PaginationRequestSchema,
+} from '../schemas/Utils';
+import fs from 'fs';
+import { pipeline } from 'stream';
+
+import util from 'node:util';
+import { randomUUID } from 'node:crypto';
+
+const pump = util.promisify(pipeline);
+
+export const listPageable = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  try {
+    const { user } = request as ISecureRequest;
+    const { page, size } = PaginationRequestSchema.parse(request.query);
+
+    const totalContentCount = (+page - 1) * +size;
+
+    const items = await prisma.item.findMany({
+      where: { campusId: user.campusId },
+      take: size,
+      skip: totalContentCount,
+      include: {
+        image: {
+          select: {
+            path: true,
+          },
+        },
+      },
+    });
+
+    const totalItemsCount = await prisma.item.count({
+      where: { campusId: user.campusId },
+    });
+
+    const totalPages = Math.ceil(totalItemsCount / size);
+
+    const response = items.map((u) =>
+      ItemResponseSchema.parse({
+        ...u,
+        image: u?.image ? u?.image.path : null,
+      }),
+    );
+
+    return reply.code(STANDARD.OK.statusCode).send({
+      totalPages,
+      totalCount: totalItemsCount,
+      content: response,
+      totalContentCount,
+      lastPage: totalPages === page,
+    });
+  } catch (err) {
+    return handleServerError(reply, err);
+  }
+};
+
+export const listById = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  try {
+    const { user } = request as ISecureRequest;
+    const id = 0;
+
+    const item = await prisma.item.findUniqueOrThrow({
+      where: { id, campusId: user.campusId },
+      include: {
+        image: {
+          select: {
+            path: true,
+          },
+        },
+      },
+    });
+
+    return reply
+      .code(STANDARD.OK.statusCode)
+      .send(ItemResponseSchema.parse(item));
+  } catch (err) {
+    return handleServerError(reply, err);
+  }
+};
+
+export const uploadItemImage = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  try {
+    const { itemId } = ItemIDRequestParamSchema.parse(request.params);
+    const imageFile = await request.file();
+    if (!imageFile) {
+      throw new AppError('É preciso passar o arquivo de imagem', 400);
+    }
+    const filename = imageFile.filename;
+    const path = `./public/images/${randomUUID()}`;
+
+    await pump(imageFile.file, fs.createWriteStream(path));
+
+    const item = await prisma.item.findUniqueOrThrow({
+      where: {
+        id: itemId,
+      },
+      include: {
+        image: true,
+      },
+    });
+
+    if (item.imageId) {
+      fs.unlink(('.' + item.image?.path) as string, async function () {
+        await prisma.image.delete({
+          where: { id: item.imageId || 0 },
+        });
+      });
+    }
+
+    const image = await prisma.image.create({
+      data: {
+        name: filename,
+        filetype: imageFile.mimetype,
+        path: path.slice(1),
+        size: imageFile.file.bytesRead,
+      },
+    });
+
+    await prisma.item.update({
+      where: {
+        id: itemId,
+      },
+      data: {
+        imageId: image.id,
+      },
+    });
+
+    return reply.code(STANDARD.OK.statusCode).send(image);
+  } catch (err) {
+    return handleServerError(reply, err);
+  }
+};
 
 export const create = async (request: FastifyRequest, reply: FastifyReply) => {
   try {
@@ -19,6 +160,14 @@ export const create = async (request: FastifyRequest, reply: FastifyReply) => {
       data: {
         ...payload,
         userId: user.id,
+        campusId: user.campusId,
+      },
+      include: {
+        image: {
+          select: {
+            path: true,
+          },
+        },
       },
     });
 
@@ -53,9 +202,48 @@ export const remove = async (request: FastifyRequest, reply: FastifyReply) => {
   try {
     const { id } = request.params as IRequestIdParamSchema;
 
+    const item = await prisma.item.findFirstOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (item.status === 'EXPIRED') {
+      throw new AppError('Você não pode remover um item expirado', 400);
+    }
+
     await prisma.item.delete({
       where: {
-        id: Number(id),
+        id,
+      },
+    });
+
+    return reply.code(STANDARD.NO_CONTENT.statusCode);
+  } catch (err) {
+    return handleServerError(reply, err);
+  }
+};
+
+export const claimItem = async (
+  request: FastifyRequest,
+  reply: FastifyReply,
+) => {
+  try {
+    const { id } = request.params as IRequestIdParamSchema;
+
+    const item = await prisma.item.findFirstOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (item.status === 'EXPIRED') {
+      throw new AppError('Você não pode remover um item expirado', 400);
+    }
+
+    await prisma.item.delete({
+      where: {
+        id,
       },
     });
